provider/user: Fix errors & split types
